+++ src/StoreValue.mjs
@@ -1,0 +1,3 @@
+/**
+ * @typedef {{module?: string | null, key: string, value: *}} StoreValue
+ */