+++ src/Value.mjs
@@ -1,0 +1,3 @@
+/**
+ * @typedef {{module: string, key: string, value: *}} Value
+ */